Pulled out CountedPingProtocol into its own file
